# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen warehouse (WMS).",
        "title": "Warehouse Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@anda.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/employees": {
            "get": {
                "description": "Mengambil daftar employee yang difilter berdasarkan warehouse_code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get Employee Berdasarkan Warehouse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode Warehouse",
                        "name": "warehouse_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Employee"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing warehouse_code query",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Membuat employee baru dengan data yang diberikan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Buat Employee Baru",
                "parameters": [
                    {
                        "description": "Data Employee Baru",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateEmployee"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/employees/by-warehouse": {
            "get": {
                "description": "Mengambil daftar employee yang difilter berdasarkan warehouse_code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get Employee Berdasarkan Warehouse",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode Warehouse",
                        "name": "warehouse_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Employee"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing warehouse_code query",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "description": "Mengambil satu data employee berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get Employee Berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Employee"
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Employee tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data employee berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Hapus Employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Employee tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Memperbarui data employee (bisa sebagian) berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Update Employee (Parsial)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update employee",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdatedEmployee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "ID atau data JSON tidak valid",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Employee tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Employee": {
            "type": "object",
            "properties": {
                "employee_code": {
                    "type": "string"
                },
                "employee_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_role": {
                    "type": "integer"
                },
                "role": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Role"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                },
                "warehouse": {
                    "$ref": "#/definitions/models.Warehouse"
                },
                "warehouse_code": {
                    "type": "string"
                }
            }
        },
        "models.Inventory": {
            "type": "object",
            "properties": {
                "code_product": {
                    "type": "string"
                },
                "code_warehouse": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_size": {
                    "type": "integer"
                },
                "product": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Product"
                        }
                    ]
                },
                "quantity": {
                    "type": "integer"
                },
                "size": {
                    "$ref": "#/definitions/models.Size"
                },
                "warehouse": {
                    "$ref": "#/definitions/models.Warehouse"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Category"
                        }
                    ]
                },
                "description_product": {
                    "description": "Diasumsikan bisa null",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_category": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "product_code": {
                    "type": "string"
                },
                "product_details": {
                    "description": "Relasi (Has Many)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductDetail"
                    }
                },
                "product_name": {
                    "type": "string"
                }
            }
        },
        "models.ProductDetail": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "code_product": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "id_size": {
                    "type": "integer"
                },
                "product": {
                    "description": "Relasi (Belongs To)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Product"
                        }
                    ]
                },
                "size": {
                    "$ref": "#/definitions/models.Size"
                }
            }
        },
        "models.Role": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "models.Size": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Warehouse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "inventories": {
                    "description": "Relasi (Has Many)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Inventory"
                    }
                },
                "location_description": {
                    "type": "string"
                },
                "warehouse_code": {
                    "type": "string"
                },
                "warehouse_name": {
                    "type": "string"
                }
            }
        },
        "request.CreateEmployee": {
            "type": "object",
            "required": [
                "employee_name",
                "id_role",
                "password",
                "warehouse_code"
            ],
            "properties": {
                "employee_name": {
                    "type": "string"
                },
                "id_role": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "warehouse_code": {
                    "type": "string"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "request.UpdatedEmployee": {
            "type": "object",
            "properties": {
                "employee_name": {
                    "type": "string"
                },
                "id_role": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "warehouse_code": {
                    "type": "string"
                }
            }
        }
    }
}